Fix MSID class
This was broken for the simple MSID('ccd_temp', '2012:001',
'2012:002') case.  In that case, the MSID init is still from MSIDset,
and passed a dictionary of slot_data.  The original logic was not
handling that object type for msid_data.  Now it is handled with an if
statement (but could certainly be improved).