Set aca_hdr3 to only read 8x8 files
Remove the data masking method and mask now by file transitions from a
file list from the archfiles db instead of by reading all of the files
and using the samples with IMGSIZE != 8 to determine transitions.

Also, as the comment says:
make a shared time ndarray that is the union of the time sets in the
slots.  The ACA L0 telemetry has the same timestamps across slots,
so the only differences here are caused by different times in
non-TRAK across the slots (usually SRCH differences at the beginning
of the observation)