Update aspect_1 status query to use science_2 info
The "P" state in the aspect_1 products appears unreliable in some
cases.  The science_2 quality (Pending, Superceded, Default,
Reprocessed?) appears to be slightly more reliable, and this query
digs through the associations in the tables to check the status of all
of the science_2 products associated with an aspect_1_id.