Reset vv status for updated asp1 products
This code change is mostly for asp_l1 products that were first
ingested as a _last version and then reingested at the default
version.  Just skipping them (not updating the processing table)
because they were "already in the db" mean that the V&V processing
left the _last directory hanging around and did not make an updated
default release directory.