Calculate residuals for omitted stars
This code creates a mock gsprops entry for stars which were manually
removed in processing or reprocessing (and hence have no gsprops entry).
With a "mock" entry, the code will continue and calculate residuals
for missing stars if there are centroids available in the
*adat files.  This is not yet implemented for removed FID lights.
Since we now have data that was unused in processing, we store a "used"
status in the saved files (json, shelve, hdf5).