Rearrange order of some V&V slot processing
This now copies the gsprops status and such over to the slot report
structure before trying to calculate the residuals.  This is related
to the 2 centroid minimum of the last change that had the side effect
that the star status pieces from gsprops were not collected before
bailing.  Tested on obsid 16064 which has one centroid for slot 7.