Improve utility of starcheck timeline lookup for corner case
By doing the db query with a sort order, for the corner case where the
date matches both the datestop of one timeline and the datestart of another,
return the timeline that matches on datestart.  Hopefully, overlaps will not
be a bigger deal than this in our use cases.  This is a fix for obsid 17587.