Only try to rerun data fetch and V&V if there's new data
This code was attempting to get new asp1 products and rerun V&V
for all obsids that had "pending" or provisional data.  These changes
set it to fetch and check the version numbers of the content it has
and the "default" content and try to only fetch and redo V&V
(or mark to have V&V redone) if the state in the database has really
changed.