Add a script to make a test starcheck database from recent data
Even on just a few weeks, this seems to take too long to be a
standard test, so I'm just putting the logic here for safekeeping.
It could also be a "long" test for ska_testr, but that wouldn't
conveniently let you check test code compatibility without a dev ska.