Replace shell "find" with os.walk
This code was set to try to find and ingest "new" starcheck.txt files using a Linux shell
"find" command set to find only files newer than the last ingested file. The time of the
last ingested file was stored by updating *another* file with that file's mod or change
time. The "find" command syntax was ugly, doing this in the shell was also not pretty, and
using a separate tracking file added a place where this could go wrote. This commit adds
the file mod time to the database (so an external tracking file isn't needed) and then
uses os.walk to get files and their mod times to include or exclude them from an update.
The os.walk code is borrowed from kadi.events.orbit_funcs .