Set the filter on new starchecks to be at or after mtime
Instead of just trying to ingest starcheck.txt files newer than the last one in the
database, try to ingest those with mod times at or after that last time.  This should be
slightly safer, as the process to update the last recorded mtime in the database happens
before the successful ingest of the file to which that time corresponds.