Only add manvrs to the list if they have content
It looks like the regex that was expanded to try to get intermediate maneuvers
is now trying to parse the star catalog as a maneuver, failing, but still
adding extra maneuver "instances" to the data structure.  This quick fix
does not fix the maneuver top level regex, but instead just stops the code
from adding manvr dicts to the list if they don't at least have a time and
a first quaternion.